Fix warning

+++ clibs/remotedebug/rdebug_stdio.cpp
@@ -22,7 +22,7 @@ static int redirect_read(lua_State* L) {
     luaL_Buffer b;
     luaL_buffinit(L, &b);
     char* buf = luaL_prepbuffsize(&b, (size_t)len);
-    size_t rc = self.read(buf, len);
+    size_t rc = self.read(buf, (size_t)len);
     if (rc == 0) {
         return 0;
     }
